+++ src/cli/commands/dump-new.ts
@@ -5,8 +5,8 @@ import { $command, $options } from '../commands';
 import type { Context } from '../../context';
 import { s } from '../../utils';
 
-import { dumps } from '../../db-v2/dumps';
-import { into } from '../../db-v2';
+import { dumps } from '../../db/dumps';
+import { into } from '../../db';
 
 const redact = (connectionString: string) => {
   const uri = new URL(connectionString);
