+++ src/db/certs.ts
@@ -1,4 +1,4 @@
-import { type Infer, t, table } from '../db-v2';
+import { type Infer, t, table } from '../db';
 
 export const certs = table('certs', {
   id: t.id(),
